requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
schema==0.7.7
sqlparse==0.5.3
urllib3==2.3.0
